+++ src/components/Custom/Common/Receipt/ReceiptDetail.tsx
