+++ src/components/Custom/Common/Receipt/ReceiptFooter.tsx
