+++ src/components/Custom/Common/Receipt/ReceiptLayout.tsx
