+++ src/components/Custom/Common/Receipt/Submitted.tsx
